add cluster component
