+++ frontend/src/routes/ClusterManagement/Clusters/components/ClusterDestroy.tsx
@@ -1,27 +1,16 @@
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { useHistory } from 'react-router-dom'
 import { AcmButton, AcmPageProcess } from '@open-cluster-management/ui-components'
 import { ExternalLinkAltIcon } from '@patternfly/react-icons'
 import { useTranslation } from 'react-i18next'
-import { ManagedClusterDefinition } from '../../../../resources/managed-cluster'
 import { NavigationPath } from '../../../../NavigationPath'
 import { launchLogs } from './HiveNotification'
+import { AddCluster } from './AddCluster'
 import { ClusterStatus, Cluster } from '../../../../lib/get-cluster'
-import { getUserAccess } from '../../../../lib/rbac-util'
 
 export function ClusterDestroy(props: { isLoading: boolean; cluster?: Cluster }) {
     const { t } = useTranslation(['cluster'])
     const history = useHistory()
-
-    const [canCreateCluster, setCanCreateCluster] = useState<boolean>(false)
-    useEffect(() => {
-        const canCreateCluster = getUserAccess('create', ManagedClusterDefinition)
-        canCreateCluster.promise
-            .then((result) => setCanCreateCluster(result.status?.allowed!))
-            .catch((err) => console.error(err))
-        return () => canCreateCluster.abort()
-    }, [])
-
     return (
         <AcmPageProcess
             isLoading={props.isLoading}
@@ -53,26 +42,7 @@ export function ClusterDestroy(props: { isLoading: boolean; cluster?: Cluster })
                     {t('button.backToClusters')}
                 </AcmButton>
             }
-            secondaryActions={
-                <>
-                    <AcmButton
-                        isDisabled={!canCreateCluster}
-                        tooltip={t('common:rbac.unauthorized')}
-                        variant="link"
-                        onClick={() => history.push(NavigationPath.createCluster)}
-                    >
-                        {t('managed.createCluster')}
-                    </AcmButton>
-                    <AcmButton
-                        isDisabled={!canCreateCluster}
-                        tooltip={t('common:rbac.unauthorized')}
-                        variant="link"
-                        onClick={() => history.push(NavigationPath.importCluster)}
-                    >
-                        {t('managed.importCluster')}
-                    </AcmButton>
-                </>
-            }
+            secondaryActions={<AddCluster type="button" buttonType="link" />}
         />
     )
 }
